refactor(portal-api): extract widget theming to services/widget_theme.py
Behavior-preserving lift of the widget theming cluster (_HEX_COLOR_RE,
_hex_to_rgb, _readable_text_color WCAG-luminance, _merge_css_variables) out of
app/api/partner.py. Pure string->dict transforms; app.api.partner re-imports
only _merge_css_variables (the widget-config / public-bot-config route handlers
call it). The hex helpers + regex are internal to the new module. Added 10 direct
characterization tests pinning the WCAG luminance branch (#2b32fd->white,
#fcaa2d->dark), dark-theme/position vars, hex validation, and the css_variables
override-merge (previously only indirectly covered for empty configs); green
before and after.

partner/widget suite: 368 passed. ruff check + format clean.
